feat: Stage 7 - Import, Media Export, Notification baseline, security hardening & v1.0.0
Real semester archive Import (defensive parse, versioned Zod validation,
allow-listed field-by-field record construction, warn->preview->confirm UX)
consuming the untrusted-input contract Export already self-validated
against. Separate, independently-versioned Media Export (personal images
only, Course/Unit/filename zip via JSZip). A real Notification engine
baseline: an in-app "starting soon" reminder banner (Tier 1) plus
best-effort local Notification scheduling while the tab is open (Tier 2),
both gated behind the existing Class reminders toggle. Security hardening
(Dependabot, npm audit in CI, documented dependency-advisory exception) and
an accessibility pass on every new flow. Closes ROADMAP.md's Stage 7 and
ships the first production-ready 1.0.0.

See docs/STAGE_5_REPORT.md for the full report.

+++ src/domain/archive.ts
@@ -93,7 +93,7 @@ const scheduleTemplateSchema = z.object({
   id: z.string(),
   courseId: z.string(),
   type: z.string(),
-  dayOfWeek: z.number().min(0).max(6),
+  dayOfWeek: z.number().int().min(0).max(6),
   startTime: z.string(),
   endTime: z.string(),
   location: z.string().optional(),
